# [.NET] Removed dependency on System.Text.Json for improved startup time