--- a/dotnet/Gherkin.sln
+++ b/dotnet/Gherkin.sln
@@ -20,6 +20,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Gherkin.Benchmarks", "Gherkin.Benchmarks\Gherkin.Benchmarks.csproj", "{4DC5C858-3F32-44E7-8FF6-7D85A16F7FF7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Gherkin.SourceGenerator", "Gherkin.SourceGenerator\Gherkin.SourceGenerator.csproj", "{0DF5A047-E6CB-44FE-9A79-AB55DF5C87D6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -38,6 +40,10 @@ Global
 		{4DC5C858-3F32-44E7-8FF6-7D85A16F7FF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4DC5C858-3F32-44E7-8FF6-7D85A16F7FF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4DC5C858-3F32-44E7-8FF6-7D85A16F7FF7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0DF5A047-E6CB-44FE-9A79-AB55DF5C87D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0DF5A047-E6CB-44FE-9A79-AB55DF5C87D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0DF5A047-E6CB-44FE-9A79-AB55DF5C87D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0DF5A047-E6CB-44FE-9A79-AB55DF5C87D6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE