--- a/dotnet/Gherkin/Gherkin.csproj
+++ b/dotnet/Gherkin/Gherkin.csproj
@@ -19,7 +19,7 @@
     <Product>Gherkin Parser</Product>
     <PackageId>Gherkin</PackageId>
     <Authors>Cucumber Ltd, Gaspar Nagy</Authors>
-    <Copyright>Copyright &#xA9; Cucumber Ltd, Gaspar Nagy</Copyright>
+    <Copyright>Copyright © Cucumber Ltd, Gaspar Nagy</Copyright>
     <Description>Cross-platform parser for the Gherkin language, used by Cucumber, SpecFlow and other Cucumber-based tools to parse feature files.</Description>
     <PackageTags>specflow gherkin cucumber</PackageTags>
     <PackageProjectUrl>https://github.com/cucumber/gherkin</PackageProjectUrl>
@@ -32,12 +32,8 @@
     <PackageOutputPath>bin/$(Configuration)/NuGet</PackageOutputPath>
   </PropertyGroup>
 
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.0'">
-    <PackageReference Include="System.Text.Json" Version="8.0.5"/>
-  </ItemGroup>
-
   <ItemGroup>
-    <EmbeddedResource Include="gherkin-languages.json"/>
+    <ProjectReference Include="..\Gherkin.SourceGenerator\Gherkin.SourceGenerator.csproj" ReferenceOutputAssembly="false" OutputItemType="Analyzer" />
   </ItemGroup>
 
   <ItemGroup>